+++ src/test/fixtures/asignatura.fixture.ts
@@ -1,12 +1,3 @@
-/**
- * Fixtures del módulo
- * 
- * Simular registros existentes en la base de datos.
- * 
- * Cuando un caso de prueba requiera un escenario diferente, se recomienda
- * crear una copia mediante el operador spread (...) y sobrescribir únicamente
- * las propiedades necesarias.
- */
 
 export const asignaturaFixture = [
     {
