+++ src/test/fixtures/carrera.fixture.ts
@@ -1,12 +1,4 @@
-/**
- * Fixtures del módulo
- * 
- * Simular registros existentes en la base de datos.
- * 
- * Cuando un caso de prueba requiera un escenario diferente, se recomienda
- * crear una copia mediante el operador spread (...) y sobrescribir únicamente
- * las propiedades necesarias.
- */
+
 
 export const carreraInformaticaFixture = {
