+++ src/test/fixtures/matricula.dto.fixture.ts
@@ -1,0 +1,9 @@
+import { EstadoOMatricula } from "../../modules/matricula/matricula-estado.enum";
+
+
+export const matriculaDTOFixture = {
+    ID_estudiante: 2,
+    ID_carrera: 1,
+    semestre: 1,
+    estado: EstadoOMatricula.ACTIVA,
+};
