+++ src/test/fixtures/plazos.fixture.ts
@@ -1,0 +1,17 @@
+
+
+export const plazoActivoFixture = {
+
+    inicio: new Date("2026-01-01"),
+
+    fin: new Date("2026-12-31"),
+
+};
+
+export const plazoVencidoFixture = {
+
+    inicio: new Date("2025-01-01"),
+
+    fin: new Date("2025-12-31"),
+
+};
